requirements.txt
aiogram==3.7.0
openai==1.35.15
grpcio>=1.50.0
grpcio-tools>=1.50.0
SQLAlchemy==2.0.20
aiosqlite==0.19.0
python-dotenv==1.0.0
APScheduler==3.10.4
aiofiles>=0.8.0
pandas>=2.0.0
requests>=2.28.0
